# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: VoiceAsset Server API
  version: 0.22.0
  summary: Public REST contract for the VoiceAsset modular monolith.
  description: >-
    VoiceAsset provides local web sessions, immutable audio ingestion, durable
    transcription jobs, encrypted workspace-scoped ASR provider profiles,
    transcript retrieval, authenticated byte-range audio delivery, and bounded
    asynchronous waveform derivatives. Resource access is always scoped to the
    authenticated workspace.
  license:
    name: GNU Affero General Public License v3.0 or later
    identifier: AGPL-3.0-or-later
servers:
  - url: /
    description: Server hosting this contract
...
security: []
paths:
...
  /api/v1/admin/jobs:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      tags: [Administration]
      summary: List durable workspace jobs
      description: >-
        Requires `admin:read`. Results are workspace-scoped and ordered by
        update time then identifier, both descending. The opaque cursor is
        bound to the workspace and complete state/kind filter set. Job payload,
        idempotency material, and worker identifiers are never returned.
      operationId: listAdminJobs
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationsListLimit'
        - $ref: '#/components/parameters/ListCursor'
        - $ref: '#/components/parameters/JobStateFilter'
        - $ref: '#/components/parameters/JobKindFilter'
      responses:
        '200':
          $ref: '#/components/responses/OperationsJobListResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
...
  /api/v1/admin/members:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      tags: [Administration]
      summary: List workspace members
      description: >-
        Requires `admin:read`. Results are workspace-scoped and ordered by
        update time then user identifier, both descending. The opaque cursor is
        bound to the workspace and complete role/status filter set. Password
        hashes and credentials are never returned. This read is audited.
      operationId: listAdminMembers
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationsListLimit'
        - $ref: '#/components/parameters/ListCursor'
        - $ref: '#/components/parameters/MemberRoleFilter'
        - $ref: '#/components/parameters/MemberStatusFilter'
      responses:
        '200':
          $ref: '#/components/responses/MembershipListResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
    post:
      tags: [Administration]
      summary: Create a workspace member
      description: >-
        Requires the Owner role and `admin:write`. Creates a local user and an
        active membership atomically. The supplied password is hashed and is
        never returned or recorded in audit metadata.
      operationId: createAdminMember
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberRequest'
      responses:
        '201':
          $ref: '#/components/responses/MembershipCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '413':
          $ref: '#/components/responses/RequestTooLargeResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
  /api/v1/admin/members/{member_id}:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/MemberID'
    patch:
      tags: [Administration]
      summary: Change a workspace member role or status
      description: >-
        Requires the Owner role and `admin:write`. Uses optimistic concurrency,
        prevents demoting or disabling the last active Owner, and revokes all
        sessions and API keys when a membership is disabled.
      operationId: updateAdminMember
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/MembershipIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
      responses:
        '200':
          $ref: '#/components/responses/MembershipResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '412':
          $ref: '#/components/responses/MembershipVersionConflictResponse'
        '413':
          $ref: '#/components/responses/RequestTooLargeResponse'
        '428':
          $ref: '#/components/responses/PreconditionRequiredResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: voiceasset_session
      description: >-
        Opaque HttpOnly session cookie. Unsafe cookie-authenticated operations
        also require Origin to equal the configured public origin.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: voiceasset_refresh
      description: >-
        Opaque HttpOnly refresh cookie scoped to `/api/v1/auth`. It is accepted
        only by the rotation endpoint together with the exact configured Origin.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VoiceAsset opaque token
      description: >-
        Opaque session token beginning with `va_`, or scoped API key beginning
        with `va_pat_`. If both cookie and bearer credentials are sent, they
        must contain the same token. API keys cannot revoke web sessions.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; oversized values are replaced.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    Origin:
      name: Origin
      in: header
      required: false
      description: >-
        Browser origin. Required and exact for unsafe cookie-authenticated
        operations; optional for bearer authentication.
      schema:
        type: string
        format: uri
...
    OperationsListLimit:
      name: limit
      in: query
      required: false
      description: Maximum number of administration read-model items to return; defaults to 50.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    JobStateFilter:
      name: state
      in: query
      required: false
      description: Exact durable job-state filter.
      schema:
        type: string
        enum: [queued, running, retry_wait, succeeded, failed, cancelled]
    JobKindFilter:
      name: kind
      in: query
      required: false
      description: Exact bounded durable job-kind filter.
      schema:
        type: string
        pattern: '^[a-z][a-z0-9_]{0,99}$'
...
    MemberRoleFilter:
      name: role
      in: query
      required: false
      description: Exact workspace-member role filter.
      schema:
        $ref: '#/components/schemas/MemberRole'
    MemberStatusFilter:
      name: status
      in: query
      required: false
      description: Exact workspace-membership status filter.
      schema:
        $ref: '#/components/schemas/MemberStatus'
    ListCursor:
      name: cursor
      in: query
      required: false
      description: Opaque continuation cursor returned by the preceding page.
      schema:
        type: string
        minLength: 1
        maxLength: 1024
...
    MemberID:
      name: member_id
      in: path
      required: true
      description: Workspace member user identifier.
      schema:
        type: string
        format: uuid
...
    MembershipIfMatch:
      name: If-Match
      in: header
      required: true
      description: Exact quoted membership version returned in ETag.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
      example: '"3"'
...
  headers:
    XAPIVersion:
      description: Stable API namespace served by this process.
      schema:
        type: string
        const: v1
    XRequestID:
      description: Caller-provided or server-generated request correlation ID.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    XServerVersion:
      description: Semantic server build version.
      schema:
        type: string
        minLength: 1
    Location:
      description: Relative URL of the created or completed resource.
      schema:
        type: string
    ETag:
      description: Quoted resource version or lowercase SHA-256 entity tag.
      schema:
        type: string
...
    WWWAuthenticate:
      description: Bearer authentication challenge.
      schema:
        type: string
        const: 'Bearer realm="voiceasset"'
...
  responses:
...
    OperationsJobListResponse:
      description: Stable page of workspace-scoped durable jobs without payload or lease-owner data.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OperationsJobList'
...
    MembershipListResponse:
      description: Stable page of workspace members without credentials or password hashes.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MembershipList'
    MembershipCreatedResponse:
      description: Local user and active workspace membership created atomically.
      headers:
        Location:
          $ref: '#/components/headers/Location'
        ETag:
          $ref: '#/components/headers/ETag'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Member'
    MembershipResponse:
      description: Current workspace membership state.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Member'
...
    BadRequestResponse:
      description: Invalid JSON, identifier, idempotency key, or operation input.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedResponse:
      description: Session credential is missing, malformed, expired, revoked, or ambiguous.
      headers:
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ForbiddenResponse:
      description: Origin validation failed or the authenticated principal lacks the required scope.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundResponse:
      description: Resource does not exist in the authenticated workspace.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowedResponse:
      description: The route exists but does not support this HTTP method.
      headers:
        Allow:
          description: Comma-separated methods accepted by the route.
          schema:
            type: string
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictResponse:
      description: Current resource, membership, idempotency input, upload state, or asset state conflicts with this request.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MembershipVersionConflictResponse:
      description: If-Match did not equal the current membership version.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RequestTooLargeResponse:
      description: JSON or binary request body exceeds the operation limit.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalErrorResponse:
      description: Request could not be completed due to an internal failure.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailableResponse:
      description: A required application service is not configured or not ready.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PreconditionRequiredResponse:
      description: A required If-Match version was omitted.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    MembershipList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Member'
        next_cursor:
          type: string
          minLength: 1
          maxLength: 1024
    Member:
      type: object
      additionalProperties: false
      required: [id, workspace_id, email, role, status, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
        role:
          $ref: '#/components/schemas/MemberRole'
        status:
          $ref: '#/components/schemas/MemberStatus'
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MemberRole:
      type: string
      enum: [owner, admin, editor, viewer, agent]
    MemberStatus:
      type: string
      enum: [active, disabled]
    CreateMemberRequest:
      type: object
      additionalProperties: false
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
        password:
          type: string
          minLength: 12
          maxLength: 1024
          writeOnly: true
        role:
          $ref: '#/components/schemas/MemberRole'
    UpdateMemberRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        role:
          $ref: '#/components/schemas/MemberRole'
        status:
          $ref: '#/components/schemas/MemberStatus'
    OperationsJobList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/OperationsJob'
        next_cursor:
          type: string
          minLength: 1
          maxLength: 1024
    OperationsJob:
      type: object
      additionalProperties: false
      required:
        - id
        - created_by
        - kind
        - state
        - attempts
        - max_attempts
        - retryable
        - available_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        kind:
          type: string
          pattern: '^[a-z][a-z0-9_]{0,99}$'
        state:
          type: string
          enum: [queued, running, retry_wait, succeeded, failed, cancelled]
        attempts:
          type: integer
          minimum: 0
        max_attempts:
          type: integer
          minimum: 1
          maximum: 20
        retryable:
          type: boolean
          description: >-
            True only when this exact failed job is currently eligible for one
            audited manual retry.
        available_at:
          type: string
          format: date-time
        lease_expires_at:
          type: string
          format: date-time
        last_error_code:
          type: string
          pattern: '^[a-z][a-z0-9_]{0,99}$'
        result_revision_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/APIError'
...